Describe/requirements.txt
pycocotools~=2.0.4
matplotlib~=3.5.2
scikit-image -
torch~=1.8.1
nltk~=3.7
torchvision~=0.9.1
numpy~=1.21.6
pillow~=9.1.1 -
tqdm~=4.64.0
jupyter==1.0.0
opencv-python==4.6.0.66
gradio==3.1.1